requirements.txt
# requirements.txt
pandas
mysql-connector-python
SQLAlchemy
chardet
kaggle
python-dotenv==1.0.1